float

+++ assets/skio-plan-picker-component.js
@@ -1007,7 +1007,7 @@ export class SkioPlanPickerComponent extends LitElement {
                             <span style="font-weight: 700;"> 
                             $${
                               (() => {
-                                                const price = (this.price(group.selected_selling_plan, false) / 100) - parseFloat(this.subscription_discount) + parseFloat(this.subscriptionPricingConfig['next_price']);
+                                const price = (this.price(group.selected_selling_plan, false) / 100) - parseInt(this.subscription_discount) + parseFloat(this.subscriptionPricingConfig['next_price']);
                                 return price % 1 === 0 ? price.toFixed(0) : price.toFixed(2);
 
                               })()
